agent/requirements.txt
google-adk
aiohttp
aiohttp_cors
aiortc
python-socketio
a2a-sdk
Pillow
